[DEV] SEOS-1326 Updated I2C components on rpi3 to instatiate all i2c busses
With this commit all available i2c busses on the rpi3 are instatiated.
The user can chose if he wants bus 0 or 1 by defining the corresponding component in the assembly section
This is simmilar to how the uart driver handles all instantiations. 

Fixed a minor bug in the I2C_INSTANCE_CONNECTION_CLIENT where the Macro instantiated multiple different connections with the same name.