[DEV] SEOS-1326 Added support for the Sabre board
Implemented new platform into the I2C driver. This Driver uses the device tree driver of the sabre board.
However the driver has some implementation errors, so currently the modified files are inside the repo

Renamed i2c.h to if_i2c.h because the previous name is used inside the libplattsupport.

Added support for the i2c_slave_init function into the rpi3 as a not implemented. The function is needed by the DTB driver.

Most of the code was taken from the i2c_demo by Lukas Graber